requirements.txt
requests==2.21.0
WTForms==2.2.1
Flask_WTF==0.14.2
Flask==1.1.1
Flask-Cors==3.0.8
python-dotenv==0.10.3
GitPython==3.0.2
PyGithub==1.43.8
uWSGI==2.0.18